% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aklcc}
\alias{aklcc}
\title{Alaska Landscape Conservation Cooperative.}
\format{A SpatialPolygonsDataFrame with 5 features. See \code{aklcc$LCC_Name}.}
\usage{
aklcc
}
\description{
A vector map of Alaska Landscape Conservation Cooperative areas.
For further details, see \url{https://lccnetwork.org}
}
\keyword{datasets}
