% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alaska}
\alias{alaska}
\title{State of Alaska.}
\format{A SpatialPolygonsDataFrame with a single feature.}
\usage{
alaska
}
\description{
A vector map of the state of Alaska boundary.
}
\keyword{datasets}
