% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cavm}
\alias{cavm}
\title{Circumpolar Arctic Vegetation (Alaska).}
\format{A SpatialPolygonsDataFrame with 3 features. See \code{cavm$Name}.}
\usage{
cavm
}
\description{
A vector map of SNAP's derived version of the Alaska region of the Circumpolar Arctic Vegetation map layer.
For further details on the source data, see \url{http://www.geobotany.uaf.edu/cavm}.
}
\keyword{datasets}
