% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecoreg}
\alias{ecoreg}
\title{Alaska ecological regions (ecoregions).}
\format{A SpatialPolygonsDataFrame with 32 features. See the \code{COMMONER}, \code{LEVEL_2}
and \code{LEVEL_1} attributes, e.g.: \code{ecoreg$Level_2}.}
\usage{
ecoreg
}
\description{
A vector map of Alaska ecoregions.
}
\keyword{datasets}
