% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fmz}
\alias{fmz}
\title{Alaska Fire Service fire management zones.}
\format{A SpatialPolygonsDataFrame with 14 features. See \code{fmz$REGION}.}
\usage{
fmz
}
\description{
A vector map of Alaska Fire Service fire management zones.
For further details, see \url{https://afs.ak.blm.gov/fire-management/zones.php}.
}
\keyword{datasets}
