% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lcc}
\alias{lcc}
\title{Alaska and Canada Landscape Conservation Cooperative.}
\format{A SpatialPolygonsDataFrame with 5 features. See \code{lcc$LCC_Name}.}
\usage{
lcc
}
\description{
A vector map of Alaska and Canada Landscape Conservation Cooperative.
For further details, see \url{https://lccnetwork.org}
}
\keyword{datasets}
