% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{snappolys}
\alias{snappolys}
\title{Basic metadata for all data sets in snappoly}
\usage{
snappolys()
}
\value{
a data frame.
}
\description{
This function returns a data frame with basic meta data for all data sets in \code{snappoly}.
This includes data object names, spatial domain, number of features, number of variables, ID column name,
whether or not the SpatialPolygonsDataFrame has multilevel/overlapping features, and a short description.
}
\examples{
snappolys()
}
