% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tpa}
\alias{tpa}
\title{Terrestrial protected areas.}
\format{A SpatialPolygonsDataFrame with 8 features. See \code{tpa$MGT_AGENCY}.}
\usage{
tpa
}
\description{
A vector map of Alaskan and Canadian terrestrial protected areas.
Management agencies include the Government of British Columbia, Parks Canada Agency,
Government of Yukon Department of Environment, US Bureau of Land Management,
US Department of Defense and US Department of Energy, US Fish and Wildlife Service, US National Park Service,
and Alaska state Department of Natural Resources.
}
\keyword{datasets}
