% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{canada}
\alias{canada}
\title{Canadian provinces.}
\format{A SpatialPolygonsDataFrame with 13 featrues. See the \code{NAME} attribute: \code{canada@data$NAME}.}
\usage{
canada
}
\description{
A vector map of Canadian provinces.
}
\keyword{datasets}
